[FIX] web_enterprise: allow to resize the window when in fullscreen editor mode
The backend editor has a button which allows to make it fullscreen. To
do that, every DOM element of the standard backend is hidden thanks to
display: none. This induced a browser crash when resizing the window
while being in this mode: the code which handles rerendering the navbar
when there are too many items in it relied on the fact the navbar is
always in a visible display... which is wrong because of this
fullscreen mode (see a3ffbfbe5a9fab2342fe40c9f69a765fd055401b).

This commit fixes the problem by simply checking the navbar is visible
before trying to reorder its items.
Note: maybe entering the fullscreen mode should be done with a more
robuste way.