[FIX] project_timesheet_synchro: Connect with standard login if no OAuth
If you connect to your OAuth 'Odoo.com' account and try to login a
database which is a new one and has no OAuth module installed. The
application will never connect to the database.

So to support both, OAuth and new 'Odoo Online' databases, through the
'Odoo.com' account for users having both type of databases. We will try
to connect with the OAuth token to the database, and if there is no
OAuth module installed, fall back to the standard login form.