[FIX] helpdesk: wrong context on dashboard sla failed
When the dashboard is displayed, two criteria are used to get the numbers
in "My Tickets" block:

- filter out tickets not assigned to me,
- filter out closed tickets,

And then depending on failed SLA's or ticket priority, tickets are
counted.

But for failed SLA's there was a disrenspency when clicking on the
filters and displaying the kanban views caused by a wrong context.

opw-697015