[FIX] web: full height Kanban view on mobile
When we display a Kanban view on a mobile device, the columns height can
be smaller than the screen height.

This leads to some problem, of navigation when youo want to swipe left
or right, you can't do it anywhere on the screen, you have to swipe over
the columns.

This problem is caused by the LESS of webclient_layout which change the
position from 'absolute' to 'static' when the 'o_view_manager_content'
is displayed on mobile. This has been done for the nested Kanban views,
that means when a 'o_view_manager_content' is in another one.

So to fix this, we will set the position of the 'o_view_manager_content'
to 'static', only when the it has also the class 'o_form_field' which
means that it's a nested view. And let the 'absolute' position when the
'o_view_manager_content' is the main one