[FIX] web_mobile: 'Add to mobile' button visible from desktop
When you are not in the mobile application, the button 'Add to Mobile'
should not be shown on the partner form view, because this button calls
a native function in the mobile application to add contact to the
phone.

However, if you are connected to your Odoo instance and go back to see
a partner through the breadcrumb, the button is displayed.

This is caused, by the method start which set the widget invisible when
it's created, so when the same widget is displayed again, it's not
invisible anymore.

To fix this, we set in the widget a variable saying if we are or not in
the mobile application. And display the widget if this variable is 'True'.