[FIX] crm_voip: tel link on mobile devices
When a field has the widget phone set on it, and if the module
'crm_voip' is installed, the tel link doesn't work anymore on mobile.

This is caused by the override of the 'phone' widget in 'crm_voip' which
try to lauch the voip module which cannot work on mobuile devices.

So to fix it, we check the screen size, before overriding the widget to
check if we keep a 'tel' link or if we override it with voip
functionnalities