[FIX] account_reports: overflow with accounting reports
The CSS rule `white-space: nowrap;` was added
to avoid line breaks within the table lines.

However, this is not mandatory for the table
headers, and this leaded to some overflow
issues in the display of the table:
an horizontal scroll bar appeared for instance
when a customer name was quite long, even
with a 1080p resolution.

opw-692521