[FIX] account_reports: write on context
- Go to Accounting > Reports > Partner Ledger
- Change the filter "Customers and suppliers" to "Customers"
- Nothing happens, the filter is reset

The method `return_context` is defined in the model
`account.report.context.common` but the field associated to the filter
(`account_type`) is defined on the model
`account.partner.ledger.context`. Therefore, `self` and `context` are
not, in this case, the same model and we need to check that the field
exists on the appropriate model.

It was working by change for other fields since they start with `add_`
or `remove_`. In this case, we don't check if they are on the model.

opw-694812