[FIX] website_version: hide new/duplicate version in translation mode
Creating a new version while being in another language
than the master language does not work,
and this is not expected to work.

Therefore, we hide the possibility to avoid
user errors.

This is not possible either to add a new page
while you are not in the master language anyway.

opw-703191