[FIX] account_reports: communication
In a follow-up report, the communication of follow-up line is defined as
the reference of the related invoice. However, in case of a migration
from another software to Odoo, there might not be any invoice linked to
the journal item.

In this commit, we use the AML name as fallbacks is the invoice
reference is empty.

opw-703379