[FIX] account_batch_deposit: create_batch_deposit new API call
The call to `create_batch_deposit` in the server action
`action_account_create_batch_deposit` has been converted
as a new API call, with the browse records set `records`,
while the method still expected the record ids to browse
in the method signature, in the kwargs `active_ids`.

As the server action is a data, for which any change
would require an update of the database,
we consider better to change the method signature rather
than the call in the server action, as this won't require
an upgrade.

Nevertheless, we are aware this is a bit risky,
in case a community / custom modules calls
the method the old way, but the probability
is rather low as 10.0 is quite recent, and this is an advanced
feature.

In case we have a report for the above,
we will then add the `active_ids` as an optional
kwargs

opw-704083
Closes #684