[FIX] web: wraparound no create editable list
In e0c4ec5 a change was done to improve the behavior in some
editable list instances, but this introduced a small issue for
list with line creation disabled.

With this change, doing tab or enter at last line should keep the
previous behavior of wrapping around and not creating a new line.

opw-703748

note: enterprise equivalent for community odoo/odoo@d40edd7052