[FIX] mrp_quality: default values on qualiy.alert
MO: manufacturing order
WO: work order

When you create a quality alert from a MO and a WO, default values for
the product product were wrong (it was the product template id that was
passed).

When you create a quality alert from the MO, the user id of the quality
alert was defined as the user id of the MO, which is wrong. It should be
the user creating the alert. To fix this one, we remove the defaut value
from the context as the default value on the creation of quality alert
will correctly fetch the current user.

When creating a quality alert from a WO, the company id of the quality
alert was defined as the current company of the user, which is wrong. It
should be defined as the company of the origin MO.

Now we have a consistent behavior of the quality alerts creation between
MO and WO.