[FIX] web: allow list view horiz. scrolling in any context
When the list view is both wide and tall, you had to scroll vertically
to reach the horizontal scrollbar and be able to scroll horizontally.
Now, the o_content element should always be in charge of scrolling
horizontally.

Note: the problem has been solved at list view level to be sure to not
break anything else. Indeed, solving the problem at o_content level
was breaking mobile view (where the o_content is still working the
same way but without being in charge of scrolling). In mobile, the
"problem" is left as it is as the user scrolls without using the
scrollbars.