[FIX] mps: when changing a 'to supply' of a product from auto to manual, it should only delete the existing to_supply of that product
Before, when you wanted to change the value for the To Supply for a period
for a certain product in the MPS report in order to give a manual value
instead of an automatic one, it would delete all existing manual values
of all products for that report.  (just try setting two manual values for the same period)

With this fix, it only deletes its own values and you
can set two manual values for different products on the same period.