[FIX] mrp_account: correct the (automatic) calculation of the cost of the finished product
When the finished product of an mo is using average or real
cost method, its price is computed automatically based on
the costs of the components and the work done.  For the components,
we should however also take care of the cost method used and not
suppose that everything can be accounted for in real price.

Before, when a component was in standard price, it would still
be accounted for like it were real price.

We also need to divide the total cost by the quantity done, not the quantity foreseen.
Before, it was done by the quantity foreseen,
so if you had an MO with a quantity of 5 and you did
post inventory after producing 2, it would divide the total
costs by 5 instead of 2.  We should also avoid to do
costing when nothing was produced.

Also a fix needed to be done in community to make sure
the consumedmoves don't contain previously done moves.
See commit c453358

A test was added to check all this by doing multiple
produces and posting the inventory in between.

commit c453358

[IMP] ensure ones

[FIX]