[FIX] account_yodlee: some account does not have itemData informations
when trying to display an account to choose from, it might happens that inside the json response, some accounts are of another type than bank and does not have an itemData.account property. This fix will prevent a crash if this is the case and will only display accounts that we can choose from.