[FIX] *_contract: stop switching environment for compute fields
Using sudo is nice and all, but if you assign the computed fields'
values in another environment than the one from the call to the compute
function, your values are never taken into account.

Use a sudoed record for computation if needed, but assign on the
record from the self environment.