[FIX] sale_contract: invoice date
When an invoice is created thanks to the scheduled task (cron) or the
"⇒ Generate Invoice" button, the invoice date is always set to the
recurring next date. This is not correct, since the invoice date should
be the date the invoice is validated.

This can lead to fiscal inconsistencies if the invoice generation is
late, and the invoice is generated with a date in the past (e.g. if the
reporting of this period is already closed).

opw-705925