[FIX] account_reports: general ledger display 80 first items
In the general ledger report, when opening an account to see
the journal items, the number of results is limited on purpose
to 80 items, order by date ASC.

To know if there are more than 80 items, the query limit is
set to 81 (to know if there is at least one more)
https://github.com/odoo/enterprise/blob/10.0/account_reports/models/account_general_ledger.py#L206-L207

Than, if there is 81 results,
the results are shrinked to 80 items.

The thing is, they were shrinked to the 80 last items,
instead of the 80 first items, giving the feeling
there was a journal item missing.

e.g.
 - for a range of one day (Jan 04), you have only 6 results (all displayed, thus),
 - then you add some days to the range (Jan 04 to Jan 06), making the
   number of results greater than 80 (and thus, not all displayed).

You were missing one of the 6 items of the one day range (Jan 04) because
the first one was popped out of the result.

This revision makes sure to choose the first 80 items instead
of the last 80 items.

opw-706248