[FIX] web_studio: edit report not only in debug
The edit button should be available in Studio even if the user is
not in debug mode. As this is not the default behaviour in the
report client action, the method that update the control panel
buttons has been overwritten.

This commit also reduces the number of records in report edition to one.