[FIX] account_reports: available companies in multicompany
In Odoo 10.0, the record rule regarding
`res.company` slightly changed: a user
in the access rights groups can now see all companies

Because of this, when a user within the access rights group
changed of company while browsing a report of another company
(e.g. display general ledger of company A, then switch to company B)
a security warning was raised.

This is because the report context was not regenerated because
the availabe companies for this user didn't change
(he can see all of them) while he can no longer access
the account.account records of the former company
(because of the multi company record rule on account.account)

The available companies of the report are no longer
the company the user can see, but the company
of the journal the user can see
(there is a multi-company record rule on these journals)

opw-673762