[FIX] helpdesk: remove overkill readonly/invisible from form
These `attrs` were overkill, and leaded to issues in some cases:

e.g. when changing the `partner_id`, the email was visibly changed,
but then at save, tt wasn't applied, because the field was set readonly,
and readonly values are not sent to  write operation by the web client.
The new behavior acts exactly as in project_issue

for the partner name, it wasn't possible to set one when there was none,
but you could bypass it anyway by setting a partner, then unsetting it,
and changing the name that was set by the partner onchange.

opw-706645