[FIX] account_reports: do not truncate in Excel export
Partner name and communication are truncated if longer than 35
characters. This can be problematic when the user wants the complete
details of an entry.

We cannot change this in stable since it will break the layout of the
reports. However, in the case of an XLSX export, this is acceptable to
keep the full length.

opw-706060