[FIX] web_studio: deal with list view without field
Before this commit, you could only see a grey screen when you
opened a list view without field.

The hook before the first column is now displayed even if there
is no column to remedy this problem.

The target of this hook was something like
<xpath expr='//field[...]' position='before'>

which cannot apply if there is no field and thus becomes
<xpath expr='//tree' position='inside'>