[FIX] account_reports: selected context
Fix 6c04ffae29 reverts what was done with fix 4ac697853ef. A consequence
is that, in some circumstances, the context selected while displaying
the report might be different from the report exported. In other words,
we export a report which is not the report displayed.

This is a quick fix made for emergency, a more consistent fix is to send
the context id when exporting instead of doing another search.

opw-707887