[FIX] web_studio: catch exception in studio view
Issue:
-----
This commit prevents a traceback caused by a customized Studio view.
The issue occured when you tried to edit a inherited view with groups on it.
When an user didn't have access to it, the content of this view wasn't
added in the arch. So even if the studio view is applied last,
it wasn't possible to apply xpath on it.

You may also have this problem if you uninstall a module.

Example:
-----
You are connected as admin user and you open Studio.
You are able to see elemens "A" and "B" because you are member
of groups "a" and "b", respectively.
You try to add a new element after the "A" element. So, an xpath will be
generated from "A".
At this step, everything works fine and you save the changes.

Now, you are logged in with another user with the group "b" only.
But the studio view doesn't find the "A" element and just crash.

Fix:
-----
To fix this, we apply each xpath from the studio view with a fallback.
In case of exception, the incident is only reported in the log, and
the inheritance proceeds with the next xpath.

We also warn users of possible problematic xpaths.
These cases are the elements added with Studio and that are located after
a node defined in a view protected by a group.

Additional Information:
-----
There are several ways to restrict a part of the view to a group of users:

(1) In the arch, you can use the "groups" attribute on a node to specify which
user groups have the right to see the content of that node.
The result will be an architecture rendered with html elements hidden
by 'oe_invisible' class.

(2) On the view, you can use the "groups_id" field to give view access
to certain users only.
The view will not be applied to the architecture if the user is not allowed
to access it. So, nothing will be rendered.

In the future:
-----
We will probably remove the "groups_id" field from the "ir.ui.view" template.
We will find a way to use the "groups" attribute on some xpath node where it was
not possible to do it.
Example:
    <xpath expr="//node" position='attributes'>
        <attribute name='attr'>value</attribute>
    </xpath>

Maybe we should split the studio view in several others:
The studio view would then be a view inherited from the view you are
trying to edit. Like we already did.
But this doesn't solve the fact that we don't know in which group we have to add
a new node that a user is trying to add.
You can find this kind of situation when you try to add a new node after an
element "A" introduced by a view protected by a group, but before an element "B"
coming from another view without group restriction.