[FIX] blackbox: remove pro forma orders once paid
To repoduce the issue:
1. create order 1 in tab 1, do not pay it
2. switch to tab 2 (order 1 saved in pro forma)
3. create and pay order 2
4. reopen and pay order 1
5. close and reopen pos window
=> order 1 is back

The reason is that, when the order 1 is paid, `order` represent the paid non-pro
forma order and o.data represent the previously saved pro forma order.
As long as the order 1 is not modified, it will be always the saved order that
is compared. `o.data.blackbox_pro_forma` will always be `=== true`.

opw-703821