[FIX] l10n_be_reports: Bad injection in SQL request
In function execute of psycopg2, the second argument is used for the sql
arguments. In this way, all the SQL arguments will be traited according
their type.

As written in the psycopg doc:

Warning Never, never, NEVER use Python string concatenation (+) or string parameters
interpolation (%) to pass variables to a SQL query string. Not even at gunpoint.

opw:709763