[FIX] website_contract: use object field instead of blind search
Changes made in the way transactions are created from an online quote
means that the transaction for the SO will not necessarily have
exactly the same name. For example, if the successful transaction
is called SO2017/0001-1 instead of SO2017/0001, the system will fail
to find the correct transaction and won't find any payment method.