[FIX] account_reports: Show all the tax in account.generic.tax.report
The active=False taxes must be displayed in the accounting reports.
The function _query_get just applied domain on account.move.line and
for example in the "Déclaration périodique de taxes", the domain can contain
('tax_ids.tags_ids', 'in', ...) where the taxes can be inactive. So for inactive
taxes, tags_ids is not reachable and the domain is then wrong.

opw:709996