[FIX] web_studio: sub-group properties not shown in sidebar
Before this commit, the click event was propagated to the parent.
So, when we clicked on a group in a group, the properties of the parent was showed
instead of the child.

A similar issue was fixed by _is_handled for pages in a group.
But it appears totally unnecessary knowing that it is possible to use "stopPropagation".