[FIX] sale_contract: Subscription sequence number consumed by analytic accounts
Since the below revision:
7134f5899d95e8624c8ce34ccc4c0166b358ce41

The `code` fields of an analytic account
has as default value a number of the sequence
`sale.subscription`.

Before this revision, this default value
was applied only if the analytic account
was linked to a `sale.subscription`,
as `sale.subscription` inherits (with s)
from `account.analytic.account`, and the default
was set in the `sale.subscription` `_defaults`
dict.

This has as side effect to consume a subscription
sequence number at each `account.analytic.account`
creation, even if the analytic account is not linked
to a `sale.subscription`.

This was not intended.

opw-708266