[FIX] account_reports: remove pointer cursor on fa-phone
Having a pointer cursor on an icon gives
the feeling to the user something can
be achieve by clicking on it.

This was never intended to trigger
an action when clicking on the phone icon
on a customer statement.

Besides, I am not able to find
a use case where this was actually useful.

If a use case is found, the pointer cursor
will have to be added with another, more specific,
class than just `i.fa`, which is way too generic.

opw-710076