[FIX] web_studio: export: module dependencies
Exported data may be based on data defined in other modules, and
those modules need to be listed in the 'depends' key of the manifest.

It worked fine, except for some char fields in 'ir.actions.act_window'
and 'ir.actions.report.xml' that implicitely refer to models, and the
modules defining those models need to be listed in 'depends' as well.