[FIX] web_studio: don't create default view at import
When creating a new model in Studio, a (simplified) default form
view is automatically created, alongside access rules.

However, when importing a data module created and exported with
Studio, those views and rules should not be created automatically
as they are most certainly already defined in the data module.

Creating a new default view would have the undesirable effect to
link that new view to the imported model, and the customized view
would not be used.