[FIX] account_reports: cash basis report
When the option "Cash Basis Method" is selected on the tax report, only
the percentage paid of the invoice is taken into account in the report.
However, when clicking on "Net Audit" or "Tax Audit", the journal items
displayed do not take into account this paid percentage. Therefore,
there is an inconsistency between the amount in the report and the
journal items.

This should be improved in master, but for stable, simple hide the
links. Context.

Backport of a5831b7aebdc84f42a316a666cc1e97dc9a37ca5

opw-709742