[FIX] web: print 2 and + backend pages firefox
When printing for example a list view with firefox we could have less page than
expected printed because of some CSS and how firefox behave when printing pages.

In 899bf698 there was already a change for google chrome and this commit
should fix it for firefox.

opw-695119

note: in saas-11 this fix should be applied to
web_enterprise/static/src/less/webclient_layout.less