[FIX] web_studio: avoid opening report without record
Before this commit, it was possible to open a report based on a model without any record on it.
The result of this was a uneditable blank page because we don't have any data to show.
So, you had to create at least one record.

Now, all reports are shown in a kanban view and when you click on it
an error is displayed if there is no record.