[FIX] web_studio: unwanted branding on xpath node
Before this fix, it was not possible to open some views in studio.
When you enter in studio, branding is activated to know groups on views.

But this branding shouldn't be added on xpath nodes like:
- <xpath expr="//group[@name='group_name']" position="inside"/>
- <group name="group_name" position="inside"/>
- <group name="group_name"/>
Note: these expressions are equivalent.

But the last one wasn't ignored and branding was added.
So it was not possible to locate a 'group' node with name="group_name" and
studio-view-group-ids attribute.

Now, we apply branding on all nodes and we remove it to locate node in arch.

Introduced by b5610120