[FIX] l10n_be_reports: when printing xml tax, should sum all lines starting with tags 48 and 46
It was the case in v8 and not anymore in v9. So if we happened to have a value in grid 48s44 and 48s46L, when printing
the xml report, we had two entries with tag 48 which is refused by financial institutions.

Backport of d4f4b4ee45ae91573fbc5da17bf97fc72f94c4c2