[FIX] web: fix export dialog design on IE
As the export dialog has always lots of content, the design was made
so that it always takes 100% of the viewport height. This allowed to
design the dialog content so that the fields panels do not overflow
the modal and are individually scrollable and so that the buttons are
vertically centered.

The problem is that this was not properly working on IE. Indeed, the
`.modal-body` element overflowed the `.modal-content` one. The solution
is in two parts:

 1) IE does not properly distribute heights when combining flex elements
    with max-height containers. The `.modal-content` has to be forced
    to `height: 100%;` for this to work. This is done in javascript as
    there is no way to detect `.modal-content` elements specific to
    export dialogs.

 2) IE does not properly render `flex: 1 0 0;` elements. The rule is
    supposed to tell the browser "render the element as if it was of
    dimension 0 and was able to grow to fill its parent container".
    Fortunatly, a workaround with `flex: 1 1 auto;` combined with
    `min-height: 0;` seems to achieve the same behavior for this
    case (by telling "render the element by shrinking or growing
    it to fill its parent container, without caring of its minimum
    size").

opw-725391