[FIX] account_bank_statement_import_ofx: TZ shift
- Import the following transaction:
  <STMTTRN>
    <TRNTYPE>XFER
    <DTPOSTED>20170307000000[+1:CET]
    <TRNAMT>-1200
    <FITID>AZERTYUIOP
    <NAME>Test Transaction
    <MEMO>Visa Card
  </STMTTRN>
- The date of the transaction is recorded on the 2017-03-06 instead of
  2017-03-07.

The library OfxParse returns a transaction date which is actually a
datetime in UTC, in this case 2017-03-06 at 23:00:00. The timezone
information is lost, and we keep only the date, which explains an
incorrect date.

We can assume that the user expects the date recorded to be the date in
his timezone, not the date in UTC.

Inpired from the existing method at
https://github.com/jseutter/ofxparse/blob/master/ofxparse/ofxparse.py

opw-726119