[FIX] account_reports: display more than 80 items
- Get an account with more than 80 entries for the current month, as
  well as entries before the current month.
- Go in "Accounting > Reports > General Ledger"
- Select the period "This Month"
- Unfold the account
- Click on "There are more than 80 items..."

Entries from current month as well as entries from previous months.

This is because the domain doesn't take into account the date range of
the report.

If we specify a domain instead of an active id, the method
`outboundLink` will call the action with the given domain. The context
of the action `action_move_line_select` will simply not apply since
there is no active id.

opw-725829