[FIX] website_sale_account_taxcloud: Update taxes on shop/payment display
PURPOSE
=======

Tribute to Ray Carnes for the complete explanation.

I think the design of this TaxCloud module is flawed.

The design stores the tax rate after it is returned from the API. I guess this is to cut down on API calls. The problems with this approach:

(1) Two Customers in the same zip code may have a different rate (that is, zip code isn't enough to determine the rate - https://www.avalara.com/learn/whitepapers/zip-codes-the-wrong-tool-for-the-job/ - that's the ENTIRE reason Companies like Avalara and TaxCloud exist - it is more complicated than a simple ZIP lookup - you need FULL ADDRESS lookup.

(2) Tax Rates Change - there are over 10,000 Tax Jurisdictions in the USA - and every few months a city, county, special district or state is changing a rate - so storing it would mean the updates from TaxCloud would never get into Odoo. Reporting this also, but it is making Odoo look bad and Collaboration Solutions feels like the first Customer to ever use the TaxCloud module.

SPECIFICATION
=============

We'd like to avoid contacting taxcloud at every change (add a product, change a price or discount, ...) as a tax cloud API call takes 1.5s. That would be a bad user experience.

To fix the issue of changing taxes, we are calling taxcloud before validation of the SALES ORDER, i.e. on the page SHOP/PAYMENT. That way, it's fast when creating records, but we do a final check before validation if necessary to be sure it's clean.