[FIX] inter_company_rules: _prepare_purchase_order_data
1) Enable the multi company and multi currency.
2) Set both of your company and your company, Chicago in the same base currency
(i tried with Euro).
3) Go into debug in both companies settings apps >  users > companies > inter
company rules > and tick the Create Sale Orders when buying to this company
and Create Purchase Orders when selling to this company options.
Select as a warehouse your companies warehouse
4) Create a pricelist in the sales app that has a currency different than the
base currency (i used USD)
5) Enter the sales app of a company and create an SO. Select your price list of a
different currency (here in USD). Select the other company as customer, validate the SO
6) Switch companies and enter the purchase app. Go into the RFQ and
notice that the PO has been created automatically.

Fix: Now the PO is created with the same currency_id of the SO

opw:727244