[IMP] account_taxcloud: Authorize transaction on invoice validation
PURPOSE
=======

All the transactions are logged on the taxcloud account. We have:

- Isolated sales order lines and invoice lines (Amount=100, Qty=1, TIC)
- All the cart on the document validation with all the related details (Qty, Subtotal, TIC, ...)

SPECIFICATION
=============

To be able to display a coherent report of all the transaction done with the taxcloud account, on the invoice validation (for example when clicking on the button validate), we send a new transaction 'Authorized' to taxcloud with all the invoice content.