[FIX] account_taxcloud: Use the original address if it can't be verified
PURPOSE
=======

Currently, we use `VerifyAddress` to transform an address into its most complete form.

This transformed address is then given as argument when we call the method `Lookup` that returns a rate.

ISSUE
=====

The call to the method `VerifyAddress` may fail, but the method `Lookup` will always return a rate, even if the exact address cannot be found.

This means the logic Odoo is using, which is:

 - (1) verify
 - (2) lookup

can lead to no rate being returned - in the case of the (1) call failing.

SPECIFICATION
=============

A call to `VerifyAddress`, even if it fails, should not prevent a call to `Lookup`. It should always allow the call to `Lookup`.