[FIX] account_reports: clean-up multi-company reports
Several issues have been raised in the past regarding incorrect report
options available when switching from one company to another.

This is due to the use of transient models for these reports. When
switching from one company to another, it might happen that an existing
report with incorrect an company selection list is selected. It prevents
the user to select the right company in the mentioned report.

The issue solves by itself when the transient models are cleaned-up.
However, it is not acceptable for the user to wait for it. Therefore, we
force the cleaning of these transient models manually when the user
switches from one company to another.

opw-707887