[FIX] account_taxcloud: Complete transactions on taxclout account
PURPOSE
=======

I have configured Taxcloud API and the tax lookup function works as expected.   When I Create a sales order odoo queries taxcloud for tax rate.  I Processed the delivery order, created an invoice and registered a payment.  None of these steps seem to send the correct API call to tax cloud to complete the transaction.

Actually it wasn't intended to manage the transactions in the taxcloud interface. But on this commit, we decided to 'Authorize' the transaction on the invoice validation: https://github.com/odoo/enterprise/commit/3e1189244f35ed0b55afd1c56e8243aa53eabe53

But it seems legit to log the transaction as captured when the invoice is set as paid.

SPECIFICATION
=============

Make a call the taxcloud method `Captured` on the invoice validation