[FIX] mrp_mps: missing access rights
When you are not the admin of the database and you are MRP manager, you
can access to the MPS report, but cannot modify values because of
missing access rights on 'sale.forecast' and 'sasle.forecast.indirect'.

To fix this we give full access rights on these models for the MRP
manager.