[FIX] account_yodlee: fix an error when cron is running in multicompany environment
If we have two company A and B which both have a yodlee account, when cron is running
it is running as administrator and the company taken was the one from the user. This
results in error for the second company as the user taken is not the correct one.