[FIX] web_studio: tree view not supported
Before this commit, tree views (not list views) were not supported.
There was a traceback when you tried to toggle Studio in hierarchy view.

This commit doesn't allow us to edit this kind of view in Studio,
but we are now able to render it correcly without error.

For information, the tree view doesn't exist anymore in saas-16.

Do not forward port after saas-15.

opw-728756