[FIX] web_studio: escape subviews when generating xpath
A view arch can be composed by some other subviews definition.
This kind of subview is typically defined in a field node.

Example:
<form>
  <field name="many2many">
    <form>
      [...]
    </form>
    <tree>
      [...]
    </tree>
  </field>
</form>

At this stage of development, Studio is not able to edit subviews.
So, we must not generate our xpath expressions from the nodes
contained in its sub-views.

Before this commit, it could happen that some xpath matched two
different nodes: one in the subview and another one in
the parent view.
In this case, the ORM only takes the first one but we sometimes
want to take the second one.

This occurred when the models of the two views had identical
field names like 'name', 'display_name', 'supplier',...

opw-728847