[FIX] web_studio: xpath expression based on 'div' tag more robust
Before this commit, the 'class' attribute was used to generate an
xpath based on 'div' tag.
The problem is that the value of this attribute is not unique,
so some xpath matched multiple nodes.
The result of this was that the element was tried to add in Studio
wasn't properly placed.

Now, we consider that we can use the 'name' attribute to create
unique 'div' nodes as we already did with 'field' or 'groups' nodes.