[FIX] l10n_be_reports: report vat listing partner without vat
If somme partners have no vat number set, we get a traceback when we try
to display the report that list all vat partners.

As the report should only display the partners having a VAT number, we
modified the SQL request to only take into account partners with VAT
number.