[FIX] web: FieldCharDomain: pass context to search_count
The user_context contains information like the lang and uid of the
user. It should be given for all RPCs. It wasn't the case for the
search_count done in FieldCharDomain.

This caused the following issue:

Suppose that the lang of the user is not english, and that there
is a filter in the SelectCreateDialog (opened when he click on
'Change selection') on a many2one field with a translated term.
As the search_read correctly sends the user context, the records
matching the domain will be correctly displayed in the dialog.
However, as the search_count doesn't send the context, the count
displayed in the form view isn't be the number of records listed
in the dialog.

opw~743245